requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pydantic
passlib[bcrypt]
python-jose[cryptography]
python-multipart
python-dotenv
httpx
pytest
pytest-asyncio
pytest-cov
alembic
bcrypt<4.1.0
pydantic[email]